rename root package to com.teamscale.upload and adjust groupId
