+++ pom.xml
@@ -4,7 +4,7 @@
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.teamscale</groupId>
+    <groupId>com.teamscale.upload</groupId>
     <artifactId>teamscale-upload</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>jar</packaging>
@@ -122,7 +122,7 @@
                 <groupId>org.graalvm.nativeimage</groupId>
                 <artifactId>native-image-maven-plugin</artifactId>
                 <configuration>
-                    <mainClass>com.cqse.teamscaleupload.TeamscaleUpload</mainClass>
+                    <mainClass>com.teamscale.upload.TeamscaleUpload</mainClass>
                     <imageName>teamscale-upload</imageName>
                     <buildArgs>
                         -H:+AddAllCharsets
