+++ src/main/java/com/teamscale/upload/autodetect_revision/EnvironmentVariableChecker.java
@@ -1,4 +1,4 @@
-package com.cqse.teamscaleupload.autodetect_revision;
+package com.teamscale.upload.autodetect_revision;
 
 import java.util.Arrays;
 import java.util.List;
