+++ src/main/java/com/teamscale/upload/autodetect_revision/SvnChecker.java
@@ -1,4 +1,4 @@
-package com.cqse.teamscaleupload.autodetect_revision;
+package com.teamscale.upload.autodetect_revision;
 
 public class SvnChecker {
