+++ src/main/java/com/teamscale/upload/resolve/FilePatternResolutionException.java
@@ -1,4 +1,4 @@
-package com.cqse.teamscaleupload.resolve;
+package com.teamscale.upload.resolve;
 
 /**
  * Thrown if resolving file patterns fails.
