+++ src/test/resources/com/teamscale/upload/test_utils/truststore.jks
